Face Aging using cGANs/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.7.1=py37_0
astor=0.7.1=py_0
blas=1.0=mkl
ca-certificates=2019.5.15=0
certifi=2019.3.9=py37_0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cycler=0.10.0=pypi_0
freetype=2.9.1=ha9979f8_1
gast=0.2.2=py_0
grpcio=1.16.1=py37h351948d_1
h5py=2.9.0=nompi_py37h3cb27cb_1102
hdf5=1.10.4=nompi_hcc15c50_1106
icc_rt=2019.0.0=h0cc432a_1
intel-openmp=2019.4=245
jpeg=9b=hb83a4c4_2
keras-applications=1.0.7=py_1
keras-base=2.2.4=py37_0
keras-gpu=2.2.4=0
keras-preprocessing=1.0.9=py_1
kiwisolver=1.1.0=pypi_0
libblas=3.8.0=8_mkl
libcblas=3.8.0=8_mkl
liblapack=3.8.0=8_mkl
libmklml=2019.0.3=0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.8.0=h1a1b453_0
libtiff=4.0.10=hb898794_2
markdown=2.6.11=py_0
matplotlib=3.1.0=pypi_0
mkl=2019.4=245
numpy=1.16.3=py37h873a0b8_0
olefile=0.46=py37_0
openssl=1.1.1c=he774522_1
pillow=6.0.0=py37hdc69c19_0
pip=19.1.1=py37_0
protobuf=3.8.0=py37he025d50_0
pyparsing=2.4.0=pypi_0
pyreadline=2.1=py37_1000
python=3.7.3=hb12ca83_0
python-dateutil=2.8.0=pypi_0
pyyaml=5.1=py37he774522_0
scipy=1.2.1=py37h29ff71c_0
setuptools=41.0.1=py37_0
six=1.12.0=py37_1000
sqlite=3.28.0=hfa6e2cd_0
tensorboard=1.13.1=py37_0
tensorflow=1.13.1=gpu_py37h83e5d6a_0
tensorflow-base=1.13.1=gpu_py37h871c8ca_0
tensorflow-estimator=1.13.0=py37h39e3cac_0
tensorflow-gpu=1.13.1=h0d30ee6_0
termcolor=1.1.0=py_2
tk=8.6.8=hfa6e2cd_0
tqdm=4.32.1=py_0
vc=14.1=h21ff451_3
vs2015_runtime=15.5.2=3
werkzeug=0.15.4=py_0
wheel=0.33.4=py37_0
wincertstore=0.2=py37_1002
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
zlib=1.2.11=h2fa13f4_1004
zstd=1.3.7=h508b16e_0
